% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Current}
\alias{Current}
\title{A PackedSpatRaster of 4 species with its projected distribution for current conditions}
\format{
\subsection{A PackedSpatRaster with 4 layer:}{

\describe{
\item{Spp1}{Predicted presence absence for species 1 in current
coditions}
\item{Spp2}{Predicted presence absence for species 1 in current
conditions}
\item{Spp3}{Predicted presence absence for species 1 in current
conditions}
\item{Spp4}{Predicted presence absence for species 1 in current
conditions}
}
}
}
\usage{
Current
}
\description{
A PackedSpatRaster of 4 species with its projected distribution for current conditions
}
\keyword{datasets}
