% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{CurrentLanduse}
\alias{CurrentLanduse}
\title{A PackedSpatRaster of the current landuse}
\format{
\subsection{A PackedSpatRaster with 1 layer:}{

\describe{
\item{Landuse}{current landuse}
}
}
}
\usage{
CurrentLanduse
}
\description{
A PackedSpatRaster of the current landuse
}
\keyword{datasets}
