% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_cells.R
\name{define_cells}
\alias{define_cells}
\title{Define Cells}
\usage{
define_cells(Rasterdomain)
}
\arguments{
\item{Rasterdomain}{A Raster object with any value in the cells that are part of the problem and NA values where the problem is not to be solved}
}
\value{
A character vector to be used in troblemaker
}
\description{
This function takes a Raster object and creates a character vector made for the troublemaker function
}
\examples{
data(Species)
library(terra)
Test <- Species[[1]] |>
terra::unwrap()

# Generate the "Problem.dat" file

define_cells(Test[[1]])

}
\author{
Derek Corcoran
}
