% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landuse_names.R
\name{landuse_names}
\alias{landuse_names}
\title{Landuse names}
\usage{
landuse_names(landuses = NULL, name = "Problem")
}
\arguments{
\item{landuses}{a vector with the names of the landuses}

\item{name}{The name of the output file}
}
\value{
A character vector to be used in troblemaker
}
\description{
This function takes a vector of landuse names and adds it to the problem in problemaker
}
\examples{

landuse_names(landuses =  c("Agriculture", "Forest", "Urban"))

# delete the file so the test on cran can pass this

file.remove("Problem.dat")


}
\author{
Derek Corcoran
}
