% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LanduseCombination.R
\name{LanduseCombination}
\alias{LanduseCombination}
\title{Calculate contiguity bonus based on similarity}
\usage{
LanduseCombination(
  Landuses,
  Filters,
  parameter = "ContiguityBonus",
  name = "Problem",
  verbose = FALSE
)
}
\arguments{
\item{Landuses}{is a character vector of the land-uses names}

\item{Filters}{Commonalities to look for in land-uses}

\item{parameter}{The name of the parameter to use}

\item{name}{The name of the output file}

\item{verbose}{Logical whether messages will be written while the
function is generating calculations, defaults to FALSE}
}
\value{
.dat file. This function is used for the side-effect of writing values to a file.
}
\description{
Calculate contiguity bonuses based on landuse similarity and writes them to a .dat file. The file
will be written to the location specified by the \code{name} argument. If the file
already exists, it will be overwritten. The file format is plain text, with each
line terminated by a newline character.
}
\examples{



LanduseCombination(Landuses = c("ForestDryPoor",
"ForestDryRich", "ForestWetPoor",
"ForestWetRich", "OpenDryPoor", "OpenDryRich", "OpenWetPoor",
"OpenWetRich"), Filters = c("Dry", "Wet"))

file.remove("Problem.dat")

}
