% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LockedInCells.R
\name{Locked_in_Cells}
\alias{Locked_in_Cells}
\title{Locked in ceks}
\usage{
Locked_in_Cells(Lockedin, name = "Problem")
}
\arguments{
\item{Lockedin}{A Raster object with a value of one in the cells that are locked in as solutions to the problem and NA values elsewhere}

\item{name}{The name of the output file}
}
\value{
.dat file. This function is used for the side-effect of writing values to a file.
}
\description{
This function takes a Raster object and identifies non NA cells and writes them to a .dat file as a Must Include. The file
will be written to the location specified by the \code{name} argument. If the file
already exists, it will be overwritten. The file format is plain text, with each
line terminated by a newline character.
}
\examples{
data(Species)
library(terra)
Test <- Species[[1]] |>
terra::unwrap()

# Generate the "Problem.dat" file

define_cells(Test[[1]])

file.remove("Problem.dat")

}
\author{
Derek Corcoran
}
