% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data.R
\docType{data}
\name{Species}
\alias{Species}
\title{A list of 4 species with its projected distribution for 4 landuses}
\format{
\subsection{A list of 4 Spatrasters with 4 layers each:}{

\describe{
\item{Species 1}{Predicted presence absence for species 1 in current,
forest, agriculture, and urban landuse}
\item{Species 2}{Predicted presence absence for species 1 in current,
forest, agriculture, and urban landuse}
\item{Species 3}{Predicted presence absence for species 1 in current,
forest, agriculture, and urban landuse}
\item{Species 4}{Predicted presence absence for species 1 in current,
forest, agriculture, and urban landuse}
}
}
}
\usage{
Species
}
\description{
A list of 4 species with its projected distribution for 4 landuses
}
\keyword{datasets}
