% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Troublemaker.R
\name{troublemaker}
\alias{troublemaker}
\title{Troublemaker}
\usage{
troublemaker(
  Rasterdomain = NULL,
  Rastercurrent = NULL,
  species_names = NULL,
  name = "Problem"
)
}
\arguments{
\item{Rasterdomain}{A Raster object with any value in the cells that are part of the problem and NA values where the problem is not to be solved}

\item{Rastercurrent}{raster object of current suitability}

\item{species_names}{a vector with the names of species}

\item{name}{The name of the output file}
}
\value{
A .dat file with the spatial problem formated for AMPL
}
\description{
This function is a metafunction with several functions inside of it it takes several spatial objects and generates a .dat file with a spatial dataset for AMPL
}
\examples{
data(Species)
data(Current)
library(terra)
Test <- Species[[1]] |>
terra::unwrap()

Current <- terra::unwrap(Current)

# Generate the "Problem.dat" file

TroublemakeR::troublemaker(Rasterdomain =Test[[1]],
Rastercurrent = Current,
species_names = c("Spp1", "Spp2", "Spp3", "Spp4"),
name = "Problem")

# delete the file so the test on cran can pass this

file.remove("Problem.dat")
}
\author{
Derek Corcoran
}
