% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_ampl_lines.R
\name{write_ampl_lines}
\alias{write_ampl_lines}
\title{Create budget}
\usage{
write_ampl_lines(line, name = "Problem")
}
\arguments{
\item{line}{line to be written to .dat file}

\item{name}{The name of the output file}
}
\description{
This function generates or appends the budget and transition cost to
a .dat file for ampl
}
\examples{

write_ampl_lines("param s:= 1")

file.remove("Problem.dat")

}
