% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landuse_names.R
\name{landuse_names}
\alias{landuse_names}
\title{Landuse names}
\usage{
landuse_names(landuses = NULL, name = "Problem")
}
\arguments{
\item{landuses}{a vector with the names of the landuses}

\item{name}{The name of the output file}
}
\value{
.dat file. This function is used for the side-effect of writing values to a file.
}
\description{
This function takes a vector of landuse names and writes them to a .dat file. The file
will be written to the location specified by the \code{name} argument. If the file
already exists, it will be overwritten. The file format is plain text, with each
line terminated by a newline character.
}
\examples{

landuse_names(landuses =  c("Agriculture", "Forest", "Urban"))

# delete the file so the test on cran can pass this

file.remove("Problem.dat")


}
\author{
Derek Corcoran
}
