% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_names.R
\name{species_names}
\alias{species_names}
\title{Species names}
\usage{
species_names(species_names = NULL, name = "Problem")
}
\arguments{
\item{species_names}{a vector with the names of species}

\item{name}{The name of the output file}
}
\value{
.dat file. This function is used for the side-effect of writing values to a file.
}
\description{
This function takes a vector of species names and writes them to a .dat file. The file
will be written to the location specified by the \code{name} argument. If the file
already exists, it will be overwritten. The file format is plain text, with each
line terminated by a newline character.
}
\examples{

species_names(species_names = c("Spp1", "Spp2"))
file.remove("Problem.dat")
}
\author{
Derek Corcoran
}
