% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_names.R
\name{species_names}
\alias{species_names}
\title{Species names}
\usage{
species_names(species_names = NULL)
}
\arguments{
\item{species_names}{a vector with the names of species}
}
\value{
A character vector to be used in troblemaker
}
\description{
This function takes a vector of species names and adds it to the problem in problemaker
}
\examples{
data(Species)
library(terra)
Test <- Species[[1]] |>
terra::unwrap()

# Generate the "Problem.dat" file

define_cells(Test[[1]])

}
\author{
Derek Corcoran
}
