% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/species_suitability.R
\name{species_suitability}
\alias{species_suitability}
\title{Calculate species suitability}
\usage{
species_suitability(Rastercurrent, species_names)
}
\arguments{
\item{Rastercurrent}{raster object of current suitability}

\item{species_names}{character vector of species names}
}
\value{
character string of species suitabilities
}
\description{
Calculate species suitability from a given raster and species names
}
\examples{
library(terra)
data(Current)
Current <- terra::unwrap(Current)
species_suitability(Rastercurrent = Current, species_names = c("Spp1", "Spp2", "Spp3", "Spp4"))

}
